requirements.txt
# EvoAgentBench Requirements
# Core dependencies for v1.0

# Data storage
duckdb>=0.9.0  # Primary database (can fallback to sqlite3 which is built-in)

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# JSON schema validation
jsonschema>=4.17.0

# Utilities
python-dateutil>=2.8.0

# GCP: Vertex AI Gemini, Search, Cloud Storage
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.14.0
google-cloud-discoveryengine>=0.16.0
# HTTP for real tool API calls
requests>=2.28.0

# For statistical analysis (M2)
scipy>=1.10.0

# For YAML parsing (benchmark tasks)
pyyaml>=6.0
